refactor: move runtime utils into separate module
