+++ packages/@lwc/features/src/flags.ts
@@ -6,42 +6,8 @@
  */
 type FeatureFlag = boolean | null;
 
-/*** DEFINE FEATURE FLAGS BELOW ***/
-
 export const ENABLE_FOO: FeatureFlag = true;
 export const ENABLE_BAR: FeatureFlag = false;
 export const ENABLE_BAZ: FeatureFlag = null;
 
-/*** DEFINE FEATURE FLAGS ABOVE ***/
-
-function getGlobalThis() {
-    if (typeof globalThis !== 'undefined') {
-        return globalThis;
-    }
-    // globalThis is not available in this env, so let's use some magic
-    // eslint-disable-next-line no-extend-native
-    Object.defineProperty(Object.prototype, '__magic__', {
-        get: function() {
-            return this;
-        },
-        configurable: true,
-    });
-    // @ts-ignore
-    // eslint-disable-next-line no-undef
-    const g = __magic__;
-    // @ts-ignore
-    delete Object.prototype.__magic__;
-    return g;
-}
-
-// Initialize the global configuration object if it isn't initialized already.
-const LWC_config = getGlobalThis().LWC_config || {};
-
-export const runtimeFlags = Object.assign({}, LWC_config.features);
-
-export function enableFeature(name: string) {
-    runtimeFlags[name] = true;
-}
-export function disableFeature(name: string) {
-    runtimeFlags[name] = false;
-}
+export { enableFeature, disableFeature, runtimeFlags } from './runtime';
